Remove unused metadata from generated json
The DependencySubmission API allows the content to be annotated with additional metadata.
This metadata is unused by GitHub, but could provide added value for reading the raw JSON
or for other consumers of the JSON file.

Removing this metadata since it is not currently useful and gets in the way of refactoring. We can re-add metadata later if desired.