Introduce Gradle-specific model for dependency graph
Instead of mapping directly into a GitHub-specific model, we now collect dependency
resolution data into a Gradle-specific model, and subsequently map this model to the
GitHub model.

This change is intended to make it easier to map from a fine-grained resolution
model to a coarser model useful to GitHub, and make this mapping clearer.

Future work could also allow the Gradle model to be serialized directly, or to
be mapped differently for other consumers.