Add task to resolve all build dependencies
The new `ForceDependencyResolutionPlugin` registers a task that will attempt
to resolve (but not download) all dependencies in all projects. This should
make it easier to wire the dependency extractor into an early point in the CI workflow.

If no other task is specified, then the build will resolve all known configurations by default.

The existing `GitHubDependencyExtractorPlugin` remains as-is, while the new
`GitHubDependencySubmissionPlugin` will apply this to the root build, and apply
the `ForceDependencyResolutionPlugin` to all builds.