Collect all plugin dependencies
Previously, we were ignoring any dependency resolution that occurred outside the
context of a Gradle Project. This meant that any dependencies resolved in the
context of a Settings script were not being captured.

By tracking the 'buildPath' and the 'projectPath' separately, we can now collect
all dependency resolution.

Due to a limitation in the events being processed, all non-project resolution will
be allocated to a single manifest for the entire build process. It may be possible
later to make this more precise, and assign resolution to a specific build process.