Allow job.correlator to be explicitly specified
Previously the plugin was inferring the values of job.id and job.correlator
from core GITHUB environment variables. While this may be useful in some cases,
there are many cases where it is insufficient.

With this change, the user of this plugin will directly provide these values
via custom environment variables.

The job.correlator value provided is now included in the generated file name.