Report dependencies by declaring project
This change aims to reduce redundant dependency reporting in repository snapshots,
with the aim to make the result work better with the GitHub Dependency Graph.
Previously, a dependency was reported for every project that referenced it directly
or transitively through a project dependency. This resulted in many dependency versions
 being reported many times within the snapshot, and these duplicates were then mirrored
in the GitHub Dependency Graph and Dependabot Security Alerts.

With this change, only external dependencies are reported, and only within the context
of the project that declares those dependencies. This should help reduce the massive
redundancy in the generated snapshot and make these files more useful.