Combine deps for build invocation into single manifest
Rather than attempt to split resolved dependencies into separate manifests
based on Gradle concepts of sub-project or configuration, we now combine all
resolved dependencies into a single GitHub Manifest.

There's no direct correlation in Gradle to the "manifest file" concept in GitHub,
and we no longer try to shoe-horn a partially correct mapping.

A later increment may be to annotate a direct dependency with some hints as to where the
dependency is declared, to assist with dependency updates. But this is unlikely to be
through a separate "manifest", but instead be via the "metadata" field.